Refactor DollySearchServiceConsumer and related classes to improve package structure and response handling #deploy-test-dolly-backend

+++ apps/dolly-backend/src/main/java/no/nav/dolly/opensearch/dto/SearchResponse.java
@@ -25,7 +25,7 @@ public class SearchResponse {
     public static class RegistreResponseStatus {
 
         private Long totalHitsBestillinger;
-        private Integer score;
+        private Float score;
         private String took;
         private Integer antall;
         private Integer side;
